Wash 2022 day 4
Proper solution using logic instead of wasteful set() nonsense.

Also, I worked out why the sorting "shortcut" didn't work as intended before, and it was because, with the input parsed like this:
2-4,6-8 = [[2,4],[6-8]]
I not only had to sort by the first item in each list, but also first reverse-sort by the second item in the list in case the ranges start on the same number.