Wash 2022 day 12
Fixed my code to use the bfs implementation already in aoc.py; optimized part 2 somewhat (runs in a bit under 0.2s now).

I could refactor part 2 so it searches from E to any 'a', with the bfs function tweaked a little and the neighbor adjacency thing flipped, and that would cut runtime down by a factor of like 4, but it's not really worth the effort.