Optimize 2022 day 14
An anon pointed out that a BFS type of search could be used to fill in part 2, which is several orders of magnitude faster than my naïve simulated implementation. I basically wrote a DFS where the queue is a set, just so the search order is randomized, which looks more interesting when rendered.