Solve 2022 day 15
00:36:19 (2020) / 03:29:19 (5401)
Whew boy, that was a tough one. This code barely even works.

My part 1 solution was just awful, I don't want to talk about it.

Part 2 was a little better (though still god-awfully slow and memory hungry), I made a list of line segments for each diamond the on x axis, de-overlapped the line segments, and recorded the coordinate of every 1-wide gap. I then repeated the same thing on the y-axis, and when one of those coordinates overlapped one of the x-axis ones, that one is our answer.