Improve 2022 day 19
I just noticed that one of the optimizations I tried that didn't seem to work, but I left in commented out, in fact had a typo that broke it.
Uncommenting it and correcting the typo improves part 2 calculation speed by a factor of like 100, down to ~3s, and helps on part 1 a little too.
*sigh*