Solve 2022 day 20
01:23:50 (2243) / 01:33:22 (1941)
I knew exactly how to do both parts immediately, but I wasted about an hour trying to debug an edge case not covered by the problem description OR the example input (rude). Specifically, when a number is rotated more places than the length of the list, it has to skip over itself when it loops back around to itself.

Part 2 was a really obvious modulo trick, which took me a few minutes of trial and error to hammer out the off-by-1s.