Solve 2022 day 21
00:13:04 (1204) / 00:32:57 (699)
Part 1 was straightforward.

Today's part 2 solution is well and truly unwashed: I made myself a manual input for the 'humn' number to see what was going on with the output and figure out where to start, but then I realized I could use that to find the answer manually, one digit at a time. So, instead of spending 10 minutes coding that, I spent 5 minutes finding the number by trial and error.

Also, fixed some edge cases in day 20.