Solve 2022 day 22 part 2
00:43:07 (871) / 05:40:10 (2233)
I tried to make a generalized solution for part 2, but realized I had no idea how to do that in a reasonable time, so I made a paper cube, and spent a lot of time staring at it in order to hardcode the edge "teleports" based on the real input's specific layout

Part 2 also requires some modification to the input, which manually defines which faces are which in a completely arbitrary order (based on what I wrote on my paper cube), and also includes a move interpreter I used to navigate by hand for debugging.

While satisfying, that was also incredibly tedious, and I definitely don't have it in me to go back and wash this one right now.