Solve 2022 day 24
01:40:36 (1889) / 01:52:47 (1770)
Lost about 20 minutes to using the wrong variable name in one spot, then another half hour or so to an off-by-1 when returning the answer in part 1. Since my result on the example was n+1, I should've just tried offsetting the return by -1 WAY earlier.

Part 2 did the same thing as part 1 except this time it was off by n+2, so not to be fooled again, I immediately submitted n-2 and it worked.

I've since cleaned the code up a little, and I think I stamped out the off-by-1s that were happening.

Think I'll add a renderer again.