Solve 2023 day 10
00:25:13 (935) / 01:29:47 (1544)
Part 1 was straightforward enough.

Part 2 was tricky, but I realized that I could take a shortcut by expanding the grid by a factor of 2, making the boundary "fence" extend to fill the gaps, and then doing a simple flood fill from a corner. Any coordinates not covered by either the flood fill or the path must be inside of the loop. Of those, we can count all coordinates left over that are also divisible by 2 to get the answer.

This needs washing, badly...