Solve 2023 day 14
00:12:38 (1711) / 01:02:01 (2447)
I started part 1 about 5 or 6 minutes late, which is a shame because my time would've been pretty good I think.
I don't actually know if my formula for finding the cycle offset after the cycle detection code is correct, I just twiddled semi-randomly until it worked on the example and tried submitting a few times, and it worked on the third try.