Solve 2023 day 17
02:25:20 (3093) / 02:33:05 (2638)
I knew it was obviously a Dijkstra's algorithm problem, but it took me a while to work out that I needed to add additional dimensions (heading and consecutive steps) other than just (x,y) to get Dijkstra to work. Then a whole bunch of time tracking down bugs. So many bugs...
This code is terrible.