Wash 2023 day 20 more
I realized that I don't actually need cycle detection code, the first cycle that activate each of the inputs for 'rx' is enough. I didn't see that pattern before because of an off-by-1 in my main loop (they were all -1, so I had to wait for a second activation and calculate the difference instead).