Solve 2016 day 14
This one probably took a little longer than it should. The first algorithm I thought of is to search backwards from 5-matches, which I believe is probably the most efficient way to do this, but the logic is a little tricky, and I made some mistakes that took a couple of hours to iron out.