Solve 2016 day 22
Well that was a bit little rough.
I can't be bothered to clean this one up.

This probably works on all AoC inputs, but it is definitely not a general solution, relying on a bunch of properties specific to the input(s), e.g. the rough size of the numbers in the "wall", and that everything not part of the wall is basically open space.