Optimize 2024 day 5
Replaced the bubblesort or whatever it was I wrote with Python's standard sort (probably Powersort or Timsort, I guess), which also significantly reduces code complexity.
I wish I could eliminate the cmp_to_key, but I can't for the life of me think of how to write that part the Python 3 way.