Fix 2024 day 13 z3 constraints
I changed x/y to a/b for clarity.

I accidentally instructed z3 to minimize the wrong variable, which turned out not to matter because the input was crafted for there to only be a single solution for each machine anyway: which is to say, asking for the minimum cost, as it does in the prompt, is nasty misdirection.

However, even after fixing this, specially crafted inputs could still give the wrong answer, which I presume is due to Z3 exploring possibilities with negative button presses. Adding an additional constraint for positive values of a/b fixes this, and this should now produce the correct answer for any possible input, e.g.:

Button A: X+16, Y+16
Button B: X+32, Y+32
Prize: X=3360, Y=3360

part1: 105
part2: 312500000105