Solve 2024 day 17
00:23:40 (1413) / 01:45:12 (1099)
Part 1 was obviously pretty straightforward.
Part 2, I tried...a lot of things. Seemed pretty obvious that everything happening was going to be in base 8, so I used that.
Eventually I noticed:
 - The output of the program was roughly based on 'a', in base 8, in reverse
- Appending more base-8 digits to 'a' generally only added more onto the left side of the output
- Using just a single digit, the program's output would map 1:1 to 'a' based on a sort of substitution map
Then I tried just doing that for the whole thing, building a longer number from that substitution map, and feeding it in. That kind of worked, but broke down after a few digits.
I figured I could try rebuilding the substitution map when it stopped being valid, but instead of that I decided to try something simpler, which was to just trial-and-error it by matching the output of the program to the end of the program. No match, increase 'a' by 1; yes match, shift left one place in base 8, and figure out what to append next. Basically the same as rebuilding the substitution map for each new digit appended, but simpler and more flexible. Surprisingly, that instantly produced a valid answer.