Solve 2024 day 24
01:12:29 (4517) / 04:26:49 (2495)
Started at about 12:37, annoyingly, so part 1 time is terrible today.

Part 2 was cool. Naturally I considered bruteforcing it, but the search space would've been something like prod(222, 221 ... 215), and sadly my PC would take a while to calculate 5 quintillion combinations, so I'd have to try something else.
I thought to try, using various libraries, making a graph and visualizing it. I found that networkx's implementation of Kamada–Kawai produced something usable, fed that into PyVis, and then spent some time tweaking it to make it pretty.
The first swap produced quite an obvious bump in the graph, so that was easy to find. The rest were more subtle, so I thought to try color-coding the graph, and that proved to be super useful, because my monkey brain is good at spotting when a color pattern has an error. It did still take a rather long time going over the graph by eye to find the remaining 3 errors, but I got there in the end.
I have a little bit of code in here that was supposed to verify when the calculation is correct, but evidently there's a bug in there somewhere, because it doesn't work at all. I tried submitting the first answer I generated after finding what I assumed to be every swap anyway, and it was correct.
Not sure if I'll go back over and try automating part 2 or not. Right now this is very decidedly only a tool to help you find the answer, not an automated solution.