Solve Hanukkah of Data
https://hanukkah.bluebird.sh/5784/
These were kinda cool. Worth having a look at if you like AoC, anyway.

Not all of these scripts produce a submittable answer, because the answer was a little fuzzy sometimes, and I just couldn't be bothered writing code to completely narrow down just a few results.

Below were my times according to the website:
Puzzle 1 was solved in 51 minutes.
Puzzle 2 was solved in 20 minutes.
Puzzle 3 was solved in 23 minutes.
Puzzle 4 was solved in 17 minutes.
Puzzle 5 was solved in 20 minutes.
Puzzle 6 was solved in 17 minutes.
Puzzle 7 was solved in 54 minutes.
Puzzle 8 was solved in 8 minutes.