Solve Everybody Codes 2025 quest 2
Global rank / Local times:
I	791 / 13m 15s 340ms
II	626 / 54m 05s 128ms
III	617 / 57m 13s 452ms

Wasted a bunch of time in part 2 to forgetting that rounding towards -∞ is not equivalent to just dropping the decimal (which IS equivalent to rounding towards 0).
Part 3 was obviously just part 2 with slightly tweaked parameters, and it took a few minutes to bruteforce. I can't be bothered to attempt to optimize it (especially considering pypy runs it in 15s anyway).