Optimize Everybody Codes 2025 quest 10 again
I rewrote the solver for part 3 to use a bitboard (the entire state densely packed in an int), out of curiosity to test performance, and as an exercise in using bitwise operators. It's only very slightly faster than the old part 3 function in CPython, though it is about 3x faster in PyPy.
On a standard input, the entire state fits in 28 bits using this method, which is kind of cool. Absolute headache to write, however.