Solve Everybody Codes 2025 quest 13
Global rank / score / time:
I	53 / 0 / 6m 58s 022ms
II	50 / 51 / 14m 05s 612ms
III	33 / 118 / 15m 06s 524ms

I expected part 3 to require a clever solution, but a quick skim of the data showed that we only had 500 lines of roughly half a million each, which is only 250m, and * 4 bytes = ~1GB of list, and that's really not enough for a modern machine to be bothered by. The same bruteforce I used for part 2 finished in about 15 seconds.