Wash Everybody Codes 2025 quest 13
DRYed the code out, and fixed the wasteful search for '1' that required looping over the array again (which I only left in because I didn't want to have to deal with debugging off-by-ones while going fast).

There's a smarter way to do this without having to build the entire array with bruteforce, but I probably can't be bothered.