Wash 2017 day 21
Spent way longer than was reasonable generalizing the `rotate` function that I'd previously hardcoded and then commented out. I'm not sure why that was as hard as I'd found it considering how similar it is to the `flip` function that I worked out surprisingly quickly.

I think I was in the process of trying to generalize `rotate` when exactly how the problem worked finally clicked for me (which took me almost an hour), and in a fit of annoyance I commented all my code to that point out and restarted, before realizing that a lot of it actually was completely valid later, while cleaning up. Sigh.