Solve Everybody Codes 2025 quest 17
Global rank / score / time:
I	143 / 0 / 1h 13m 34s 974ms
II	135 / 0 / 1h 20m 43s 518ms
III	93 / 58 / 4h 38m 35s 970ms

I started 59 minutes late. Had part 2 done in 21m45s local time.

Part 3 was tough, and it took me a very long time to finally think up a working algorithm that runs reasonably quickly. It still takes ~50s to finish in both pypy and CPython 3.12.

I first thought to try drawing a line on the left, bottom, and right of the volcano, and then for each radius, pathfind from S to every point on the left line, then to every point on the bottom line, then to every point on the right line, then back to S. But, that's all possible radii * half board width * half board height * half board width again =  ~75^4 full Dijkstra (wbfs=weighted bfs) over the whole board.

Then I thought, I could draw a diagonal line from the bottom 2 corners to the volcano, then try pathfinding through those 2 instead, and that would 'only' be 75^3=400k wbfses, which is maybe doable.

While that was running (it took about an hour) I tried thinking of how to optimize it more, and then an obviously better solution occurred to me, which is for every radius, to block off the right side, and pathfind to every point on the vertical line down from the volcano; then block off the left side, and pathfind back. Now we're only in the realm of 75^2, yippee.

However, that was still really slow, so it occurred to me to, rather than loop over every radius, it might be possible to iteratively refine the search. So, we start off at radius 0, and then do the same thing as before, find a loop through every point on the bottom-vertical line; of those, we take the lowest-cost loop, to establish a baseline of what the fastest possible path is. Then, if we divide that cost by 30, that equals the radius the volcano would grow to during that path. So, we expand the volcano to that radius, and check if any points in the shortest path we just found were destroyed; if there's no intersection, that's our answer. If not, pop all the 'destroyed' points out of the global grid, and restart at the last radius.

So that's the algorithm I'm committing, which takes about 8 'full' pathfinds (start to all ~75 vertical line points, then each of those back to start = 150) = ~1200 wbfses to get the answer. It's not blazing fast, but it's not too bad.

I'll fix this up later, including allowing arbitrary starting positions (it's hard coded to assume S is straight above the volcano).