Optimize Everybody Codes 2025 quest 17
Massively cut runtime by remembering that it's possible to 'batch' Dijkstra runs by starting from one point and looking for numerous points, and then "resuming" all of those points in a second search back back to the starting point by feeding the existing costs back in.
New runtime in CPython is ~2.4s and pypy is ~1.4s. 🎉