Solve Everybody Codes 2025 quest 19
Global rank / score / time:
I	73 / 0 / 33m 44s 427ms
II	76 / 25 / 59m 19s 912ms
III	84 / 67 / 2h 17m 20s 059ms

For parts 1 and 2, I used a naive Dijkstra's.
I had to rewrite a lot to handle the multiple openings per column, annoyingly.
For part 3, I reused my part 2 code but with a greatly optimized neighbors function, but did so in such a way it completely negated the advantage of using Dijkstra's in the first place. It takes a little over a minute to run.
Very messy as usual, there's a great deal of cleaning and optimization to be done here.