Solve 2025 day 3
Day   -Part 1-   -Part 2-
  3   00:13:41   01:41:35

I completely failed to work out the correct algorithm for part 2 multiple times. See the shameful procession of p2()s.
First I tried just culling 1s starting from left to right, then 2s, etc, until there's only 12 number left, and that worked sometimes, but definitely not reliably.
Next I figured a search algorithm could probably solve it pretty quickly, so I implemented a dfs (and a bfs and Dijkstra's), and the bfs technically worked, but was way too inefficient to solve the real input. So, I tried to see if I could torture Dijkstra's into doing what I needed but faster, and that was a complete failure.
While pondering what to do, one of the downsides of the search algorithm approach was that I had to re-run it for every index of the input string, and take the max output of any run. So I figured, maybe I could figure out what the starting digit had to be...it had to be the highest digit with at least 11 more numbers to the right of it, right?
So with that optimization, I went back to unsuccessfully torturing Dijkstra's for a little bit, before realizing I could just repeat the exact same process for every following digit. The second digit had to be the highest digit after the first digit with at least 10 numbers to the right of it; the third the highest with 9 to the right, etc...
That approach finishes basically instantly, and produces the correct answer.

I should've spent more time working out how I'd solve this problem on pen & paper, instead of waffling around with stupid algorithms for so long.