Wash 2025 day 9
An anon on /g/ mentioned 'coordinate compression', so I googled it to find out what it was, then made a rough guess regarding how it might be applied.
Amazingly it worked first try. New runtime is ~1-2 seconds instead of ~15 minutes. Incredible.
I was going to rewrite it to use ranges, but this might be faster (and waaaay less work).