"Solve" 2025 day 12
Day   -Part 1-   -Part 2-
 12   02:17:54   02:18:08

I wasted 2 hours studying dancing links, but before I started on that properly, I decided to write some naive code to filter out all the lines that obviously could not fit (raw tile count > rectangle). I ran it and noticed that most of the lines didn't pass, so I just tried submitting that number, and it was right.
It's a lot of effort to not fill this commit comment with every manner of swear word.